+++ src/components/private-route/private-route.tsx
@@ -6,6 +6,6 @@ type PrivateRouteProps = {
 }
 
 export function PrivateRoute({ children }: PrivateRouteProps): JSX.Element {
-  const isAuthorized = useAppSelector((state) => state.authorizationStatus);
+  const isAuthorized = useAppSelector((state) => state.user.authorizationStatus);
   return isAuthorized ? children : <Navigate to="/login" />;
 }
