+++ src/store/reducers/reducer.ts
@@ -1,0 +1,8 @@
+import { combineReducers } from '@reduxjs/toolkit';
+import { userReducer } from './userReducer';
+import { offerReducer } from './offerReducer';
+
+export const reducer = combineReducers({
+  offer: offerReducer,
+  user: userReducer
+});
